% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-directory-if-it-doesnt-exist.R
\name{create_directory_if_it_doesnt_exist}
\alias{create_directory_if_it_doesnt_exist}
\title{Create a directory if it doesn't exist}
\usage{
create_directory_if_it_doesnt_exist(dir_path)
}
\arguments{
\item{dir_path}{Location of the directory to be created}
}
\description{
All intermediate directories will be created if necessary. That is, the
directory will be created recursively
}
