% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review-rf.R
\docType{data}
\name{review_rf}
\alias{review_rf}
\title{Random forest generated by model training}
\format{An object of class \code{randomForest} of length 18.}
\usage{
review_rf
}
\description{
This is the object used to predict novel text.
}
\keyword{datasets}
