% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment}
\alias{sentiment}
\title{Classify a review as good or bad}
\usage{
sentiment(x)
}
\arguments{
\item{x}{Text to be classified, ideally a one-sentence product review.}
}
\description{
Classify a review as good or bad
}
