% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\docType{data}
\name{tfidf}
\alias{tfidf}
\title{Term frequency inverse document frequency object created in model training}
\format{An object of class \code{TfIdf} (inherits from \code{mlapiTransformation}, \code{mlapiBase}, \code{R6}) of length 5.}
\usage{
tfidf
}
\description{
This object from the text2vec package is used to map novel text to a
document term matrix with the same weighting as that used in training
}
\keyword{datasets}
