% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectoriser.R
\docType{data}
\name{vectoriser}
\alias{vectoriser}
\title{Vectoriser created in model training}
\format{An object of class \code{function} of length 1.}
\usage{
vectoriser
}
\description{
This object from the text2vec package is used to map novel text to a
document term matrix with the same terms as the document term matrix used
in training
}
\keyword{datasets}
